style: move `utils` to `common`
