+++ src/common/adapters/genres.adapter.ts
@@ -1,7 +1,6 @@
 import { Genres } from '../dtos'
 import { SpotifyArtist } from '../types/spotify'
-
-import { getMostFrequentItems } from '~/utils'
+import { getMostFrequentItems } from '../utils'
 
 export const adaptGenres = (
   artists: SpotifyArtist[],
