+++ src/common/utils/apply-authorization-header.spec.ts
