+++ src/common/utils/apply-authorization-header.ts
