+++ src/common/utils/catch-spotify-error.spec.ts
