+++ src/common/utils/catch-spotify-error.ts
