+++ src/modules/statistics/statistics.service.spec.ts
@@ -20,7 +20,7 @@ import {
   spotifyResponseWithOffsetMockFactory,
   spotifyResponseWithCursorsMockFactory,
 } from '@common/mocks'
-import { axiosResponseMockFactory } from '~/utils'
+import { axiosResponseMockFactory } from '@common/utils'
 
 describe('StatisticsService', () => {
   let statisticsService: StatisticsService
